# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_AhmedHossam777_go-mongo_internal_dto.CreateCourseDto:
    properties:
      course_name:
        maxLength: 100
        minLength: 2
        type: string
      instructor_id:
        type: string
      price:
        type: integer
    required:
    - course_name
    - instructor_id
    - price
    type: object
...
  github_com_AhmedHossam777_go-mongo_internal_dto.UpdateCourseDto:
    properties:
      course_name:
        maxLength: 100
        minLength: 2
        type: string
      price:
        minimum: 0
        type: integer
    type: object
...
  github_com_AhmedHossam777_go-mongo_internal_models.Course:
    properties:
      course_name:
        type: string
      id:
        description: If ID is empty, don't include it in JSON/BSON
        type: string
      instructor_id:
        type: string
      price:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: A RESTful API for managing courses, users, and authentication built
    with Go and MongoDB
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go-MongoDB Course API
  version: "1.0"
paths:
...
  /courses:
    get:
      description: Get a paginated list of all courses
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of courses
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all courses
      tags:
      - courses
    post:
      consumes:
      - application/json
      description: Create a new course (requires authentication)
      parameters:
      - description: Course details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_dto.CreateCourseDto'
      produces:
      - application/json
      responses:
        "201":
          description: Course created successfully
          schema:
            $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_models.Course'
        "400":
          description: Bad request - validation error or duplicate course
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new course
      tags:
      - courses
  /courses/{id}:
    delete:
      description: Delete a course by ID (requires authentication)
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Course deleted successfully
        "400":
          description: Bad request - invalid course ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Course not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete course
      tags:
      - courses
    get:
      description: Get a specific course by its ID
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Course details
          schema:
            $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_models.Course'
        "400":
          description: Bad request - invalid course ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Course not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get course by ID
      tags:
      - courses
    patch:
      consumes:
      - application/json
      description: Update course details by ID (requires authentication)
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated course details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_dto.UpdateCourseDto'
      produces:
      - application/json
      responses:
        "200":
          description: Course updated successfully
          schema:
            $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_models.Course'
        "400":
          description: Bad request - validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update course
      tags:
      - courses
  /courses/drop:
    delete:
      description: Delete all courses from the database (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: All courses deleted successfully
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Drop course collection
      tags:
      - courses
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"